requirements.txt
Flask==2.0.1
Pillow==9.5.0
opencv-python==4.7.0.72
pytesseract==0.3.10
qrcode==7.3.1
python-dotenv==1.0.0
pytest==7.3.1
numpy==1.24.3